reorganize tests folder
